// frontend/ticketing_app_ui/src/app/services/models/page-ticket-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PageableObject } from '../models/pageable-object';
import { SortObject } from '../models/sort-object';
import { TicketResponseDto } from '../models/ticket-response-dto';
export interface PageTicketResponseDto {
  content?: Array<TicketResponseDto>;
  empty?: boolean;
  first?: boolean;
  last?: boolean;
  number?: number;
  numberOfElements?: number;
  pageable?: PageableObject;
  size?: number;
  sort?: Array<SortObject>;
  totalElements?: number;
  totalPages?: number;
}